changed fonts and sizes to be relative to <body> element
It's much, much simpler, and also more in keeping with the principles
of "responsive design," set the font family and size once, and then
have all other elements inherit or adjust from this default setting. So
I deleted repeat references to the font family wherever they were the
same as the default. And for the body element, which adjusts somewhat
based on media queries, I changed font size units to em.