shorter line lengths for note area; adjusted position of #note-file-list
Most typographers recommend line lengths between 45 and 90 characters,
or between 2 and 3 *alphabets*. This commit sets #note-style to a max-
width of 40em, which works out almost exactly to 3 alphabets at the
widest. So it's still on the wider end of that recommended range. But
because the text is narrower now, users won't have to jump so far when
they move their eyes across the page as they read.

This necessitated adjusting the position of #note-file-list to maintain
the look of a paperclip attached to the notes page. However, it's
possible that the file list could also look good if it were floated to
the right of the note area, forming a third column for sufficiently
large client displays. Maybe?