converted most CSS lengths to *relative* units.
Using relative units (like em and percentages) instead of absolute
units (like pixels and inches) is central to "responsive design." I
think, if you can help it, usually the only element that needs fixed
units should be the <body> font-size, and that should probably be
adjusted based on display size using media queries. Another exception
would be bitmap graphics, which don't scale as well. CSS lengths based
on em will scale with the font size for different display devices, or
scale with user browser stylesheets which change the default font size.
And percentages will scale with the device display size.

So that's what I've done here.

Lengths of '0' are the same regardless of the unit, but in those cases
I removed 'px' anyway because it might encourage other contributors to
keep the pixel unit if they change that particular CSS property, which
generally I don't think we want.

...By the way, what does the #drag style, in `note.styl`, apply to? I
didn't change the lengths there to relative units because I couldn't
figure out what it was; I could not make this element id appear
anywhere in the markup in my test Cozy Notes app.

I also did *not* edit the "Datepicker for Bootstrap" styles in
`CNeditor.css`, though it was a little hard to tell where exactly the
Bootstrap styles ended... If those styles are from Bootstrap, I figured
you'd want to keep them the same as the upstream version. Right?

I'd like if the .Th, .To, .Th, .Lh, .Lo, and .Lu styles -- the notes
bullet points -- were indented a little more, but there would need to
be fewer indentation levels, or else the largest indents would be wider
than the #note-style container.